+++ java/effective-java/note-app/memo/src/main/java/me/staek/memo/view/decorator/WordWrapDecorator.java
@@ -1,7 +1,7 @@
 package me.staek.memo.view.decorator;
 
-import me.staek.memo.Format;
-import me.staek.memo.WordWrap;
+import me.staek.memo.dto.Format;
+import me.staek.memo.dto.WordWrap;
 
 import javax.swing.*;
 import java.util.Optional;
